requirements.txt
astroid==2.2.5
Click==7.0
Flask==1.0.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
python-ldap==3.2.0
six==1.12.0
typed-ast==1.4.0
uWSGI==2.0.18
Werkzeug==0.15.4
wrapt==1.11.2
